feat(events): add Schduler#waitAsyncEventsFinished that waits for async event execution (#2)
Closes #251
