+++ src/main/java/be/seeseemelk/mockbukkit/scheduler/BukkitSchedulerMock.java
@@ -1,21 +1,15 @@
 package be.seeseemelk.mockbukkit.scheduler;
 
 import java.util.ArrayList;
-import java.util.concurrent.ConcurrentHashMap;
+import java.util.concurrent.*;
 import java.util.List;
 import java.util.Map;
-import java.util.concurrent.Callable;
-import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Executors;
-import java.util.concurrent.Future;
-import java.util.concurrent.SynchronousQueue;
-import java.util.concurrent.ThreadPoolExecutor;
-import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicReference;
 import java.util.function.Consumer;
 import java.util.logging.Logger;
 
 import be.seeseemelk.mockbukkit.MockBukkit;
+import be.seeseemelk.mockbukkit.ThreadAccessException;
 import be.seeseemelk.mockbukkit.UnimplementedOperationException;
 
 import org.apache.commons.lang.Validate;
@@ -35,6 +29,7 @@ public class BukkitSchedulerMock implements BukkitScheduler
 	        60L, TimeUnit.SECONDS,
 	        new SynchronousQueue<>());
 	private final ExecutorService asyncEventExecutor = Executors.newCachedThreadPool();
+	private final List<Future<?>> queuedAsyncEvents = new ArrayList<>();
 	private final TaskList scheduledTasks = new TaskList();
 	private final AtomicReference<Exception> asyncException = new AtomicReference<>();
 	private long currentTick = 0;
@@ -74,7 +69,9 @@ public class BukkitSchedulerMock implements BukkitScheduler
 	public @NotNull Future<?> executeAsyncEvent(Event event)
 	{
 		Validate.notNull(event, "Cannot schedule an Event that is null!");
-		return asyncEventExecutor.submit(() -> MockBukkit.getMock().getPluginManager().callEvent(event));
+		Future<?> future = asyncEventExecutor.submit(() -> MockBukkit.getMock().getPluginManager().callEvent(event));
+		queuedAsyncEvents.add(future);
+		return future;
 	}
 
 	/**
@@ -190,6 +187,29 @@ public class BukkitSchedulerMock implements BukkitScheduler
 					}
 				}
 				pool.shutdownNow();
+			}
+		}
+	}
+
+	public void waitAsyncEventsFinished()
+	{
+		for (Future<?> futureEvent : List.copyOf(queuedAsyncEvents))
+		{
+			if (futureEvent.isDone())
+			{
+				queuedAsyncEvents.remove(futureEvent);
+			}
+			else
+			{
+				try
+				{
+					queuedAsyncEvents.remove(futureEvent);
+					futureEvent.get();
+				}
+				catch (InterruptedException | ExecutionException e)
+				{
+					throw new ThreadAccessException("Failed to wait for async events: " + e.getMessage(), e);
+				}
 			}
 		}
 	}
